+++ packages/web3-net/types/tests/network-test.ts
@@ -19,10 +19,17 @@
 
 import { Network } from 'web3-net';
 
+// $ExpectType Network
+const network_empty = new Network();
+
+// $ExpectType Network
 const network = new Network('http://localhost:5000');
 
 // $ExpectType provider
 network.currentProvider;
+
+// $ExpectType any
+Network.givenProvider;
 
 // $ExpectType any
 network.givenProvider;