+++ packages/web3-shh/types/tests/shh-test.ts
@@ -19,10 +19,17 @@
 
 import { Info, Notification, Shh } from 'web3-shh';
 
+// $ExpectType Shh
+const shh_empty = new Shh();
+
+// $ExpectType Shh
 const shh = new Shh('https://localhost:5000');
 
 // $ExpectType provider
 shh.currentProvider;
+
+// $ExpectType any
+Shh.givenProvider;
 
 // $ExpectType any
 shh.givenProvider;
@@ -32,6 +39,9 @@ shh.setProvider('https://localhost:2100');
 
 // $ExpectType BatchRequest
 new shh.BatchRequest();
+
+// $ExpectType any
+shh.extend({property: 'test', methods: [{name: 'method', call: 'method'}]});
 
 // $ExpectType Promise<number>
 shh.net.getId();