+++ CHANGELOG.md
@@ -46,8 +46,6 @@ Released with 1.0.0-beta.37 code base.
 - ``requestOptions`` added to ``WebsocketProvider`` (#2979)
 - Node >= v8.0.0 support (#2938)
 
-## [Unreleased]
-
 ## [1.2.2]
 
 ### Added
@@ -85,9 +83,36 @@ Released with 1.0.0-beta.37 code base.
 
 ## [1.2.3]
 
-### Added
+### Fixed
+
+- Fix perfect gas usage causes tx to error (#3175)
+- Fix regenerator runtime error in web3.min.js (#3155)
+- Fix TS types for eth.subscribe syncing, newBlockHeaders, pendingTransactions (#3159)
+- Improve web3-eth-abi decodeParameters error message (#3134)
+
+## [1.2.4]
 
 ### Fixed
 
-- Fix TS types for eth.subscribe syncing, newBlockHeaders, pendingTransactions (#3159)
-- Improve web3-eth-abi decodeParameters error message (#3134)
+- Fix npm installation error for scrypt-shim and websocket (#3210)
+
+## [Unreleased]
+
+## [1.2.5]
+
+### Added
+
+- ``eth_requestAccounts`` as ``requestAccounts`` added to web3-eth package (#3219)
+- ``sha3Raw`` and ``soliditySha3Raw`` added to web3-utils package (#3226)
+- ``eth_getProof`` as ``getProof`` added to web3-eth package (#3220)
+- ``BN`` and ``BigNumber`` objects are now supported by the ``abi.encodeParameter(s)`` method (#3238) 
+- ``getPendingTransactions`` added to web3-eth package (#3239)
+
+### Changed
+
+- ``eth-lib`` dependency updated (0.2.7 => ^0.2.8) (#3242)
+
+### Fixed
+
+- ``defaultBlock`` property handling fixed (#3247)
+- ``clearSubscriptions`` does no longer throw an error if no running subscriptions do exist (#3246)